chore: deleted unnecessary create group form screen files
