+++ packages/app/screens/feed/create_group_form/CreateGroupStack.js
@@ -2,7 +2,7 @@ import React from "react";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import { GroupProvider } from "@app/utils/GroupContext";
 import InitialCreateGroupScreen from "./InitialCreateGroupScreen";
-import CreateGroupFormScreen from "./CreateGroupFormScreen1";
+import CreateGroupFormScreen from "./CreateGroupFormScreen";
 
 const Stack = createNativeStackNavigator();
